+++ ResourceGenerator/M365DSCResourceGenerator.psm1
@@ -1242,6 +1242,7 @@ function Get-TypeProperties
                 if(-not [String]::IsNullOrWhiteSpace($property.Annotation.String))
                 {
                     $description =$property.Annotation.String.replace('"',"'")
+                    $description =$property.Annotation.String -replace '[^\p{L}\p{Nd}/(/}/_ -.,=:)'']', ''
                 }
                 $myProperty.Add('Description', $description)
 
@@ -1420,6 +1421,7 @@ function Get-Microsoft365DSCModuleCimClass
         $ResourceName
     )
 
+    import-module Microsoft365DSC -Force
     $modulePath = Split-Path -Path (get-module microsoft365dsc).Path
     $resourcesPath = "$modulePath\DSCResources\*\*.mof"
     $resources = (Get-ChildItem $resourcesPath).FullName
@@ -1516,9 +1518,13 @@ function Get-ComplexTypeConstructorToString
     $tempPropertyName=$returnPropertyName
 
     $valuePrefix = "getValue."
+    $referencePrefix = "getValue."
     if($isNested)
     {
-        $valuePrefix = "`$current$propertyName."
+        #$valuePrefix = "`$current$propertyName."
+        $valuePrefix = "$ParentPropertyValuePath"
+        $referencePrefix = "$ParentPropertyValuePath"
+
     }
 
     $loopPropertyName= $Property.Name
@@ -1531,6 +1537,8 @@ function Get-ComplexTypeConstructorToString
         $loopPropertyName=Get-StringFirstCharacterToLower -Value $Property.Name
         $propertyName = Get-StringFirstCharacterToLower -Value $Property.Name
         $valuePrefix += "AdditionalProperties."
+        $referencePrefix += "AdditionalProperties."
+        $referencePrefix += "$propertyName."
     }
 
 
@@ -1575,15 +1583,16 @@ function Get-ComplexTypeConstructorToString
         if($isNested -and -not $Property.IsArray)
         {
             $propRoot=$ParentPropertyName.replace("my","")
-            $valuePrefix = "current$propRoot."
+            #$valuePrefix = "current$propRoot."
+            $valuePrefix = "$referencePrefix"
 
-            $recallProperty=$propertyName
+            #$recallProperty=$propertyName
+            $recallProperty=''
             if($isParentfromAdditionalProperties)
             {
-                $recallProperty=Get-StringFirstCharacterToLower -Value $recallProperty
+                $recallProperty=Get-StringFirstCharacterToLower -Value $propertyName
             }
             $valuePrefix += "$recallProperty."
-
         }
         $AssignedPropertyName = $nestedProperty.Name
         if($nestedProperty.IsRootProperty -eq $false)
@@ -1600,7 +1609,7 @@ function Get-ComplexTypeConstructorToString
         {
             $AssignedPropertyName="'$AssignedPropertyName'"
         }
-        if((-not $isNested) -and (-not $Property.IsArray))
+        if((-not $isNested) -and (-not $Property.IsArray) -and ([String]::IsNullOrWhiteSpace($ParentPropertyValuePath)))
         {
             $valuePrefix += "$propertyName."
         }
@@ -1618,7 +1627,7 @@ function Get-ComplexTypeConstructorToString
                     -IndentCount $IndentCount `
                     -IsNested $true `
                     -ParentPropertyName $tempPropertyName `
-                    -ParentPropertyValuePath $valuePrefix `
+                    -ParentPropertyValuePath $referencePrefix `
                     -IsParentFromAdditionalProperties (-not $Property.IsRootProperty)
 
                 $complexString.append( $nestedString ) | Out-Null
@@ -1661,11 +1670,25 @@ function Get-ComplexTypeConstructorToString
 
                 if($nestedProperty.IsEnumType)
                 {
-                    $complexString.appendLine( $spacing + "if (`$null -ne `$$valuePrefix$AssignedPropertyName)" ) | Out-Null
+                    if($isNested)
+                    {
+                        $complexString.appendLine( $spacing + "if (`$null -ne `$$valuePrefix$AssignedPropertyName)" ) | Out-Null
+                    }
+                    else
+                    {
+                        $complexString.appendLine( $spacing + "if (`$null -ne `$$referencePrefix$AssignedPropertyName)" ) | Out-Null
+                    }
                     $complexString.appendLine( $spacing + "{" ) | Out-Null
                     $IndentCount ++
                     $spacing = $indent * $IndentCount
-                    $complexString.append( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$valuePrefix$AssignedPropertyName.toString()" ) | Out-Null
+                    if($isNested)
+                    {
+                        $complexString.append( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$valuePrefix$AssignedPropertyName.toString()" ) | Out-Null
+                    }
+                    else
+                    {
+                        $complexString.append( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$referencePrefix$AssignedPropertyName.toString()" ) | Out-Null
+                    }
                     $complexString.append( ")`r`n" ) | Out-Null
                     $IndentCount --
                     $spacing = $indent * $IndentCount
@@ -1673,7 +1696,14 @@ function Get-ComplexTypeConstructorToString
                 }
                 else
                 {
-                    $complexString.appendLine( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$valuePrefix$AssignedPropertyName)" ) | Out-Null
+                    if($isNested)
+                    {
+                        $complexString.appendLine( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$valuePrefix$AssignedPropertyName)" ) | Out-Null
+                    }
+                    else
+                    {
+                        $complexString.appendLine( $spacing + "`$$tempPropertyName.Add('" +  $nestedPropertyName + "', `$$referencePrefix$AssignedPropertyName)" ) | Out-Null
+                    }
                 }
 
             }
@@ -1813,7 +1843,7 @@ function Get-TimeTypeConstructorToString
         $IsNested=$false
     )
 
-    $dateString = [System.Text.StringBuilder]::New()
+    $timeString = [System.Text.StringBuilder]::New()
     $indent="    "
     $spacing = $indent * $IndentCount
 
@@ -1927,8 +1957,9 @@ function Get-ParameterBlockInformation
         $isMandatory = $false
         # Replace this one with the proper mandatory key value
         $cmdletParameter = $DefaultParameterSetProperties | Where-Object -FilterScript { $_.Name -eq $property.Name }
-        if ($null -ne $cmdletParameter `
-                -and $cmdletParameter.IsMandatory -eq $true)
+        if (($null -ne $cmdletParameter `
+                -and $cmdletParameter.IsMandatory -eq $true) `
+            -or $property.Name -eq 'Id')
         {
             $isMandatory = $true
             $parameterAttribute = "[Parameter(Mandatory = `$true)]"
@@ -2219,7 +2250,6 @@ function Get-M365DSCFakeValues
 
     foreach ($parameter in $parameters)
     {
-        $hashValue = $null
         $parameterName = $parameter.Name
         if($parameter.Name -eq "@odata.type" -and $IsGetTargetResource)
         {
@@ -2227,7 +2257,7 @@ function Get-M365DSCFakeValues
         }
         if($parameter.IsComplexType)
         {
-            $hashValue = @{}
+            [hashtable]$hashValue = @{}
             $propertyType = $workload + $parameter.Type
             if($IsGetTargetResource)
             {
@@ -2257,6 +2287,7 @@ function Get-M365DSCFakeValues
             {
                 '*.String'
                 {
+                    [String]$hashValue = ''
                     $fakeValue = 'FakeStringValue'
                     if ($parameter.Members)
                     {
@@ -2282,54 +2313,48 @@ function Get-M365DSCFakeValues
                         $fakeValue1 = $parameter.Members[0]
                         $fakeValue2 = $parameter.Members[1]
                     }
+                    [Array]$hashValue = @($fakeValue1, $fakeValue2)
                     if ($IntroduceDrift)
                     {
                         $hashValue = @($fakeValue1)
-                    }
-                    else
-                    {
-                        $hashValue = @($fakeValue1, $fakeValue2)
                     }
                     break
                 }
                 '*.Int32'
                 {
+                    [Int32]$hashValue = 25
                     if ($IntroduceDrift)
                     {
                         $hashValue = 7
-                    }
-                    else
-                    {
-                        $hashValue = 25
                     }
                     break
                 }
                 '*.Boolean'
                 {
+                    [Boolean]$hashValue = $true
                     if ($IntroduceDrift)
                     {
                         $hashValue = $false
-                    }
-                    else
-                    {
-                        $hashValue = $true
                     }
                     break
                 }
                 '*.DateTime'
                 {
+                    [String]$hashValue = ''
                     $fakeValue = ([DateTime]"2023-01-01T00:00:00").toString("$DateFormat")
                     $hashValue = $fakeValue
                     break
                 }
                 '*.DateTimeOffset'
                 {
+                    [String]$hashValue = ''
                     $fakeValue = ([DateTimeOffset]"2023-01-01T00:00:00").toString("$DateFormat")
                     $hashValue = $fakeValue
                     break
                 }
                 '*.Time*'
                 {
+                    [String]$hashValue = ''
                     $fakeValue = [Datetime]::Parse("00:00:00").TimeOfDay.toString()
                     $hashValue = $fakeValue
                     break
@@ -2427,6 +2452,10 @@ function Get-M365DSCHashAsString
                 $line = "$Space$extraSpace$key = "
                 if ($Values.$Key.isArray)
                 {
+                    if ($Values.$Key.CIMType)
+                    {
+                        $line += "[CimInstance[]]"
+                    }
                     $line += "@(`r$space    "
                     $extraSpace = '    '
                 }
@@ -2678,7 +2707,12 @@ function New-M365SchemaPropertySet
                     $mySet = $mySet.Substring(0, $mySet.Length - 1)
                     $propertySet = ", ValueMap{$mySet}, Values{$mySet}"
                 }
-                $schemaProperties += "    [Write, Description(`"$($_.Description)`")$propertySet] $($propertyType) $($_.Name)"
+                $permission = "Write"
+                if ($_.Name -eq "Id")
+                {
+                    $permission = "Key"
+                }
+                $schemaProperties += "    [$permission, Description(`"$($_.Description)`")$propertySet] $($propertyType) $($_.Name)"
                 if ($_.IsArray)
                 {
                     $schemaProperties += '[]'
@@ -2782,8 +2816,10 @@ function New-M365DSCExampleFile
     )
 
     $exportPath = Join-Path -Path $env:temp -ChildPath $ResourceName
-    Export-M365DSCConfiguration -Credential $Credential `
-        -Components $ResourceName -Path $exportPath `
+    Export-M365DSCConfiguration `
+        -Credential $Credential `
+        -Components $ResourceName `
+        -Path $exportPath `
         -FileName "$ResourceName.ps1" `
         -ConfigurationName 'Example' | Out-Null
 
@@ -2794,8 +2830,8 @@ function New-M365DSCExampleFile
     $start = $exportContent.IndexOf("{", $start) + 1
     $exampleContent = $exportContent.Substring($start, $end-$start)
 
-    $exampleFileFullPath = "$ExampleFilePath\$ResourceName\1-$ResourceName-Example.psm1"
-    $folderPath = "$ExampleFilePath\$ResourceName"
+    $exampleFileFullPath = "$Path\$ResourceName\1-$ResourceName-Example.psm1"
+    $folderPath = "$Path\$ResourceName"
     New-Item $folderPath -ItemType Directory -Force | Out-Null
     $templatePath = '.\Example.Template.ps1'
     Copy-Item -Path $templatePath -Destination $exampleFileFullPath -Force
@@ -3153,6 +3189,10 @@ function Get-ParameterBlockStringForModule
                 $parameterBlockOutput += '        [System.Boolean'
             }
             elseif ($propertyType.ToLower() -like 'system.date*')
+            {
+                $parameterBlockOutput += '        [System.String'
+            }
+            elseif ($propertyType.ToLower() -like 'system.binary*')
             {
                 $parameterBlockOutput += '        [System.String'
             }