+++ CHANGELOG.md
@@ -7,6 +7,13 @@
     FIXES [#2977](https://github.com/microsoft/Microsoft365DSC/issues/2977)
 * MISC
   * Added QA test that checks for existence of an example for each resource.
+* DRG
+  * Various fixes:
+    Remove invalid character from description in schema.mof
+    Add Id as key in schema.mof
+    Fix issue with nested CIM array from test unit
+    Remove Select-MgProfile from module
+    Fix issue with AdditionalProperties for nested objects from module
 
 # 1.23.315.2